requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
absl-py==2.3.1
    # via
    #   keras
    #   tensorboard
    #   tensorflow
astunparse==1.6.3
    # via tensorflow
beautifulsoup4==4.14.2
    # via gdown
blinker==1.9.0
    # via flask
cachetools==6.2.1
    # via google-auth
certifi==2025.10.5
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.2.0
    # via
    #   flask
    #   google-images-search
colorama==0.4.6
    # via
    #   click
    #   google-images-search
    #   tqdm
deepface==0.0.95
    # via -r requirements.in
filelock==3.20.0
    # via gdown
fire==0.7.1
    # via deepface
flask==3.1.2
    # via
    #   deepface
    #   flask-cors
flask-cors==6.0.1
    # via deepface
flatbuffers==25.9.23
    # via tensorflow
gast==0.6.0
    # via tensorflow
gdown==5.2.0
    # via
    #   deepface
    #   retina-face
google-api-core==2.27.0
    # via google-api-python-client
google-api-python-client==2.48.0
    # via google-images-search
google-auth==2.41.1
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-images-search==1.4.7
    # via -r requirements.in
google-pasta==0.2.0
    # via tensorflow
googleapis-common-protos==1.71.0
    # via google-api-core
grpcio==1.76.0
    # via
    #   tensorboard
    #   tensorflow
gunicorn==23.0.0
    # via deepface
h5py==3.15.1
    # via
    #   keras
    #   tensorflow
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.11
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
joblib==1.5.2
    # via
    #   mtcnn
    #   scikit-learn
keras==3.11.3
    # via
    #   deepface
    #   tensorflow
libclang==18.1.1
    # via tensorflow
lz4==4.4.4
    # via mtcnn
markdown==3.9
    # via tensorboard
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
ml-dtypes==0.5.3
    # via
    #   keras
    #   tensorflow
mtcnn==1.0.0
    # via deepface
namex==0.1.0
    # via keras
numpy==2.2.6
    # via
    #   -r requirements.in
    #   deepface
    #   h5py
    #   keras
    #   ml-dtypes
    #   opencv-python
    #   pandas
    #   retina-face
    #   scikit-learn
    #   scipy
    #   tensorboard
    #   tensorflow
opencv-python==4.12.0.88
    # via
    #   -r requirements.in
    #   deepface
    #   retina-face
opt-einsum==3.4.0
    # via tensorflow
optree==0.17.0
    # via keras
packaging==25.0
    # via
    #   gunicorn
    #   keras
    #   tensorboard
    #   tensorflow
pandas==2.3.3
    # via deepface
pillow==12.0.0
    # via
    #   deepface
    #   google-images-search
    #   python-resize-image
    #   retina-face
    #   tensorboard
proto-plus==1.26.1
    # via google-api-core
protobuf==6.33.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
    #   tensorboard
    #   tensorflow
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pyfiglet==0.8.post1
    # via google-images-search
pygments==2.19.2
    # via rich
pyparsing==3.2.5
    # via httplib2
pysocks==1.7.1
    # via requests
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via -r requirements.in
python-resize-image==1.1.20
    # via google-images-search
pytz==2025.2
    # via pandas
requests[socks]==2.32.5
    # via
    #   deepface
    #   gdown
    #   google-api-core
    #   google-images-search
    #   python-resize-image
    #   tensorflow
retina-face==0.0.17
    # via deepface
rich==14.2.0
    # via keras
rsa==4.9.1
    # via google-auth
scikit-learn==1.7.2
    # via -r requirements.in
scipy==1.15.3
    # via scikit-learn
six==1.17.0
    # via
    #   astunparse
    #   google-images-search
    #   google-pasta
    #   python-dateutil
    #   tensorflow
soupsieve==2.8
    # via beautifulsoup4
tensorboard==2.20.0
    # via tensorflow
tensorboard-data-server==0.7.2
    # via tensorboard
tensorflow==2.20.0
    # via
    #   deepface
    #   retina-face
    #   tf-keras
termcolor==1.1.0
    # via
    #   fire
    #   google-images-search
    #   tensorflow
tf-keras==2.20.1
    # via -r requirements.in
threadpoolctl==3.6.0
    # via scikit-learn
tqdm==4.67.1
    # via
    #   deepface
    #   gdown
typing-extensions==4.15.0
    # via
    #   beautifulsoup4
    #   grpcio
    #   optree
    #   tensorflow
tzdata==2025.2
    # via pandas
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
werkzeug==3.1.3
    # via
    #   flask
    #   flask-cors
    #   tensorboard
wheel==0.45.1
    # via astunparse
windows-curses==2.4.1 ; sys_platform == "win32"
    # via -r requirements.in
wrapt==2.0.0
    # via tensorflow
yt-dlp==2025.10.22
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
